+++ DerpNES/Instruction.cs
@@ -1,4 +1,3 @@
 ﻿namespace DerpNES;
 
-// todo: add name
-internal record struct Instruction( string Name, uint Opcode, Func<uint> Operate, Func<uint> AddressMode, uint Cycles );
+internal record struct Instruction( string Name, u8 Opcode, Func<u8> Operate, Func<u8> AddressMode, u8 Cycles );
